Add config settings for interpreter heap and stack space.
Add two configuration settings that adjust the default amount
of heap and per-thread stack space for the JVM created to run
a Rascal interpreter. Both settings can be either null or
numeric. Numeric settings are interpreted as number of MB. The
settings correspond to the -Xmx (for heap) and -Xss (for stack)
flags passed to the JVM.