Load TPLs for selected modules only.
Only load TPLs for modules that contain the name under the cursor. This
prevents type-checking or loading any modules that will not play a role
in the renaming anyway, at the cost of parsing all modules (which is
fast) and looking for specific names (which is also relatively fast).
Since type-checking is relatively slow, this should improve performance
greatly, especially for large projects and rare names.