Lifted commands in quickfixes and codeActionsProviders to real CodeActions.
Commands were the simplest way of creating this interaction, but
the LSP provides more configurability and smoother UI interactions
via the concept of a "CodeAction". A CodeAction optionally wraps
a Command and/or a DocumentEdit. The edit is always executed
first and may involve preview(s) of the different alternative
commands. The command is just the command to execute when the option
is chosen, just as before.

Command titles are lifted to CodeAction titles if none are present.

CodeActions may be "categorized" via the CodeActionKind enum. This
enum is open for users to extend. The current definition lists the
currently known list in lsp4j.

If a command calls a IDEServices.registerTextEdits, then this has
the same effect on the files and editor content as providing the edits
in the CodeAction immediately. However, UI interactions like previews
and OK/Cancel pop-ups may be different.